fix: retain prompt surface attribution through socket coordinator
